Update to C file

+++ 0x02-functions_nested_loops/101-natural.c
@@ -11,7 +11,7 @@ long int natural_35(void);
 
 int main(void)
 {
-	long int natural_35();
+	int natural_35();
 	return (0);
 }
